v0.3

+++ src/main/java/p3project/repositories/LogRepository.java
@@ -3,8 +3,7 @@ package p3project.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import p3project.classes.Eventlog;
-import p3project.classes.Changelog;
 
-public interface LogRepository extends JpaRepository<Eventlog, Integer>{
+public interface LogRepository extends JpaRepository<Eventlog, Long>{
 
 }
